gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.121.3",
    "faststream[nats]>=0.6.3",
    "pydantic-settings>=2.12.0",
    "rich>=14.2.0",
    "shared-messages",
    "strawberry-graphql[fastapi]>=0.287.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = ["ruff>=0.14.6"]

[tool.uv.sources]
shared-messages = { path = "../shared", editable = true }
